requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
ffmpeg-python==0.2.0
flower==2.0.1
future==1.0.0
gunicorn==23.0.0
humanize==4.12.3
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.3
packaging==25.0
pillow==11.2.1
prometheus_client==0.22.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
